delete Payment and PaymentLinksForSearch classes
