+++ src/test/java/uk/gov/pay/api/it/PaymentsResourceCreateIT.java
@@ -369,7 +369,7 @@ public class PaymentsResourceCreateIT extends PaymentResourceITestBase {
 
         String responseBody = postPaymentResponse(SUCCESS_PAYLOAD)
                 .statusCode(HttpStatus.SC_CREATED)
-                .contentType(JSON).log().body()
+                .contentType(JSON)
                 .header(HttpHeaders.LOCATION, is(paymentLocationFor(configuration.getBaseUrl(), CHARGE_ID)))
                 .body("payment_id", is(CHARGE_ID))
                 .body("amount", is(9999999))
